Move `SimpleUtExecutionInstrumentation` to a separate file
